requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic
python-dotenv
Pillow